requirements.txt
streamlit
langchain_google_genai 
pathway
json
requests
python-dotenv

# for map
folium
opencage